Move Bat audio helpers
